deny unsafe_code in zCore
